micro/requirements.txt
amqp==2.5.2
anyjson==0.3.3
arrow==0.15.4
attrs==19.3.0
base58==1.0.3
billiard==3.6.1.0
bitcoin==1.1.39
bitpay==2.5.1906.1
blockcypher==1.0.75
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
click==6.6
Django==2.2.3
django-mailgun==0.9.1
docker-py==1.8.0
ecdsa==0.13.2
entrypoints==0.3
flake8==3.7.9
funcsigs==1.0.2
future==0.18.2
idna==2.8
importlib-metadata==1.3.0
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
kombu==4.6.7
mccabe==0.6.1
mnemonic==0.13
more-itertools==8.0.2
numpy==1.19.4
opencv-python==4.4.0.46
packaging==19.2
path==13.1.0
path.py==12.4.0
pbkdf2==1.3
pexpect==4.7.0
Pillow==8.0.1
pluggy==0.13.1
polib==1.1.0
protobuf==3.0.0a3
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pyaes==1.6.1
pycodestyle==2.5.0
pycryptodome==3.9.4
pyflakes==2.1.1
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
python-dateutil==2.8.1
pytz==2019.1
pywallet==0.1.0
PyYAML==5.2
raven==6.10.0
redis==3.5.3
requests==2.22.0
sha256==0.1
six==1.12.0
sqlparse==0.3.0
tabulate==0.8.6
two1==3.10.9
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.8
websocket-client==0.57.0
zipp==0.6.0
